+++ src/size.h
@@ -12,13 +12,6 @@ r_obj* vec_check_recycle(r_obj* x,
                          struct vctrs_arg* x_arg,
                          struct r_lazy call);
 
-r_ssize vec_check_recyclable(
-  r_obj* x,
-  r_ssize size,
-  struct vctrs_arg* x_arg,
-  struct r_lazy call
-);
-
 static inline
 r_obj* vec_recycle(r_obj* x,
                    r_ssize size) {
@@ -31,13 +24,6 @@ r_obj* vec_recycle_fallback(r_obj* x,
                             struct r_lazy call);
 
 r_obj* list_sizes(r_obj* x, const struct vec_error_opts* opts);
-
-bool list_all_size(
-  r_obj* xs,
-  r_ssize size,
-  struct vctrs_arg* p_xs_arg,
-  struct r_lazy call
-);
 
 r_ssize df_size(r_obj* x);
 r_ssize df_raw_size(r_obj* x);